+++ components/index.ts
@@ -1,4 +1,4 @@
 export * from './Header';
-export * from './LanguageSwitch2';
+export * from './LanguageSwitch';
 export * from './SnackbarMessage';
 export * from './Layout';
